move TransactionLog to domain
