+++ src/test/scala/io/iohk/ethereum/network/p2p/validators/BlockValidatorSpec.scala
@@ -3,7 +3,7 @@ package io.iohk.ethereum.network.p2p.validators
 import akka.util.ByteString
 import io.iohk.ethereum.domain._
 import io.iohk.ethereum.network.p2p.messages.PV62._
-import io.iohk.ethereum.network.p2p.messages.PV63.{Receipt, TransactionLog}
+import io.iohk.ethereum.network.p2p.messages.PV63.Receipt
 import io.iohk.ethereum.network.p2p.validators.BlockValidator.{BlockLogBloomError, BlockOmmersHashError, BlockReceiptsHashError, BlockTransactionsHashError}
 import org.scalatest.{FlatSpec, Matchers}
 import org.spongycastle.util.encoders.Hex
